test-requirements.txt
freezegun
odoo_test_helper

odoo14-addon-shopfloor_base @git+https://github.com/OCA/wms@refs/pull/316/head#subdirectory=setup/shopfloor_base
odoo14-addon-shopfloor_mobile_base @git+https://github.com/OCA/wms@refs/pull/316/head#subdirectory=setup/shopfloor_mobile_base
odoo14-addon-shopfloor @git+https://github.com/OCA/wms@refs/pull/316/head#subdirectory=setup/shopfloor
odoo14-addon-shopfloor_mobile @git+https://github.com/OCA/wms@refs/pull/316/head#subdirectory=setup/shopfloor_mobile
